==> Python Notebooks/2 Preprocessing.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

import pandas as pd
import spacy

#Spacy library for stopwords
nlp = spacy.load('en_core_web_sm')

## TRAIN DATASET

train_df=pd.read_csv('/content/drive/MyDrive/DISSERTATION/Dataset/Extracted Dataset (csv)/train_plos.csv')

train_df

### Abstract

#Replace the last newline character in abstract
train_df['abstract']=train_df['abstract'].str.replace('\n','')

train_df.iloc[0]['abstract']

#Remove stopwords from the abstract
def rem_stopW(text):
  doc=nlp(text)
  new_text=' '.join([token.text for token in doc if not token.is_stop])
  return new_text

train_df['abstract']=train_df['abstract'].apply(rem_stopW)


### Plain Language Summary (PLS)

#Replace the last newline character in PLS
train_df['plain language summary']=train_df['plain language summary'].str.replace('\n','')

train_df.iloc[1]['plain language summary']

#Converting to CSV
train_df.to_csv('/content/drive/MyDrive/DISSERTATION/Dataset/Preprocessed Dataset/train_plos.csv')

## VALIDATION DATASET

val_df=pd.read_csv('/content/drive/MyDrive/DISSERTATION/Dataset/Extracted Dataset (csv)/dev_plos.csv')

val_df

### Abstract

#Replace the last newline character in abstract
val_df['abstract']=val_df['abstract'].str.replace('\n','')

val_df.iloc[0]['abstract']

#Remove stopwords from the abstract
def rem_stopW(text):
  doc=nlp(text)
  new_text=' '.join([token.text for token in doc if not token.is_stop])
  return new_text

val_df['abstract']=val_df['abstract'].apply(rem_stopW)


### Plain Language Summary (PLS)

#Replace the last newline character in PLS
val_df['plain language summary']=val_df['plain language summary'].str.replace('\n','')

val_df.iloc[1]['plain language summary']

#Converting to CSV
val_df.to_csv('/content/drive/MyDrive/DISSERTATION/Dataset/Preprocessed Dataset/dev_plos.csv')

## TEST DATASET


test_df=pd.read_csv('/content/drive/MyDrive/DISSERTATION/Dataset/Extracted Dataset (csv)/test_plos.csv')

test_df

### Abstract

#Replace the last newline character in abstract
test_df['abstract']=test_df['abstract'].str.replace('\n','')

test_df.iloc[0]['abstract']

#Remove stopwords from the abstract
def rem_stopW(text):
  doc=nlp(text)
  new_text=' '.join([token.text for token in doc if not token.is_stop])
  return new_text

test_df['abstract']=test_df['abstract'].apply(rem_stopW)


### Plain Language Summary (PLS)

#Replace the last newline character in PLS
test_df['plain language summary']=test_df['plain language summary'].str.replace('\n','')

test_df.iloc[1]['plain language summary']

#Converting to CSV
test_df.to_csv('/content/drive/MyDrive/DISSERTATION/Dataset/Preprocessed Dataset/test_plos.csv')
