==> Python Notebooks/3 PEGASUS.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

!pip install transformers datasets torch

!pip install transformers[torch]

!pip install accelerate -U

import pandas as pd
from datasets import Dataset
import torch
from transformers import PegasusTokenizer, PegasusForConditionalGeneration, Trainer, TrainingArguments, DataCollatorForSeq2Seq
from transformers import Seq2SeqTrainingArguments, Seq2SeqTrainer

train_df=pd.read_csv('/content/drive/MyDrive/DISSERTATION/Dataset/Preprocessed Dataset/train_plos.csv')
val_df=pd.read_csv('/content/drive/MyDrive/DISSERTATION/Dataset/Preprocessed Dataset/dev_plos.csv')

train_df.drop(columns=['Unnamed: 0'],inplace=True)
val_df.drop(columns=['Unnamed: 0'],inplace=True)

#Randomly sampling 10000 training sets from train_df
train_df = train_df.sample(n=10000, random_state=42)

train_dataset = Dataset.from_pandas(train_df)
valid_dataset = Dataset.from_pandas(val_df)

#Pegasus Model
model_name = "google/pegasus-xsum"
tokenizer = PegasusTokenizer.from_pretrained(model_name)
model = PegasusForConditionalGeneration.from_pretrained(model_name)

#Tokenizing train and validation set with Pegasus tokenizer
def tokenize_function(examples):
    inputs = examples['abstract']
    targets = examples['plain language summary']
    model_inputs = tokenizer(inputs, max_length=512, truncation=True)
    with tokenizer.as_target_tokenizer():
        labels = tokenizer(targets, max_length=512, truncation=True)
    model_inputs["labels"] = labels["input_ids"]
    return model_inputs

tokenized_train_dataset = train_dataset.map(tokenize_function, batched=True)
tokenized_valid_dataset = valid_dataset.map(tokenize_function, batched=True)

#Training Parameters
training_args = Seq2SeqTrainingArguments(
    output_dir="./results",
    eval_strategy="epoch",
    learning_rate=2e-5,
    per_device_train_batch_size=4,
    per_device_eval_batch_size=4,
    warmup_steps=500,
    weight_decay=0.01,
    save_total_limit=3,
    num_train_epochs=3,             # Number of training epochs
    save_strategy="epoch",
    predict_with_generate=True,
    gradient_accumulation_steps=2
)

data_collator = DataCollatorForSeq2Seq(tokenizer, model=model)

#Seq2Seq Training Function
trainer = Seq2SeqTrainer(
    model=model,
    args=training_args,
    train_dataset=tokenized_train_dataset,
    eval_dataset=tokenized_valid_dataset,
    tokenizer=tokenizer,
    data_collator=data_collator
)


#Model Training
trainer.train()

#Saving the trained model
trainer.save_model('/content/drive/MyDrive/DISSERTATION/Pegasus Model/v4')

## Model Testing

test_df=pd.read_csv('/content/drive/MyDrive/DISSERTATION/Dataset/Preprocessed Dataset/test_plos.csv')

test_df.drop(columns=['Unnamed: 0'],inplace=True)

test_dataset = Dataset.from_pandas(test_df)

#Tokenize test dataset
def tokenize_function(example):
    abstract = example['abstract']
    inputs = tokenizer(abstract, max_length=512, truncation=True, return_tensors="pt")
    return inputs

tokenized_test_dataset = test_dataset.map(tokenize_function)

device = torch.device("cuda" if torch.cuda.is_available() else "cpu")


generated_summaries = []

for batch in tokenized_test_dataset:
    # Convert to PyTorch tensors
    input_ids = torch.tensor(batch["input_ids"]).to(device)
    attention_mask = torch.tensor(batch["attention_mask"]).to(device)

    # Generate summary for the batch
    summaries = model.generate(input_ids=input_ids,
                               attention_mask=attention_mask,
                               max_length=512,
                               num_beams=4,
                               early_stopping=True)

    # Decode the summaries generated
    for summary_id in summaries:
        generated_summary = tokenizer.decode(summary_id, skip_special_tokens=True)
        generated_summaries.append(generated_summary)

test_df['generated_summary'] = generated_summaries

# Save to CSV file
test_df.to_csv('/content/drive/MyDrive/DISSERTATION/Dataset/test_peg_v4.csv', index=False)
