==> Python Notebooks/6 Relevance Metric.ipynb <==
# ROUGE and BERTScore

from google.colab import drive
drive.mount('/content/drive')

import pandas as pd

## ROUGE

pip install pandas rouge-score

from rouge_score import rouge_scorer

### Pegasus

peg_df=pd.read_csv("/content/drive/MyDrive/DISSERTATION/Dataset/test_peg_v4.csv")

scorer = rouge_scorer.RougeScorer(['rouge1', 'rouge2', 'rougeL'], use_stemmer=True)

#Function to calculate ROUGE scores
def calculate_rouge_scores(target, generated):
    scores = scorer.score(target, generated)
    return {
        'rouge1_precision': scores['rouge1'].precision,
        'rouge1_recall': scores['rouge1'].recall,
        'rouge1_fmeasure': scores['rouge1'].fmeasure,
        'rouge2_precision': scores['rouge2'].precision,
        'rouge2_recall': scores['rouge2'].recall,
        'rouge2_fmeasure': scores['rouge2'].fmeasure,
        'rougeL_precision': scores['rougeL'].precision,
        'rougeL_recall': scores['rougeL'].recall,
        'rougeL_fmeasure': scores['rougeL'].fmeasure
    }

rouge_scores = peg_df.apply(lambda row: calculate_rouge_scores(row['plain language summary'], row['generated_summary']), axis=1)

rouge_df = pd.DataFrame(rouge_scores.tolist())

result_df = pd.concat([peg_df, rouge_df], axis=1)

average_rouge_scores = rouge_df.mean()

#Average ROUGE scores
print("Average ROUGE Scores:")
print(f"ROUGE-1 Precision: {average_rouge_scores['rouge1_precision']:.4f}")
print(f"ROUGE-1 Recall: {average_rouge_scores['rouge1_recall']:.4f}")
print(f"ROUGE-1 F-Measure: {average_rouge_scores['rouge1_fmeasure']:.4f}")
print(f"ROUGE-2 Precision: {average_rouge_scores['rouge2_precision']:.4f}")
print(f"ROUGE-2 Recall: {average_rouge_scores['rouge2_recall']:.4f}")
print(f"ROUGE-2 F-Measure: {average_rouge_scores['rouge2_fmeasure']:.4f}")
print(f"ROUGE-L Precision: {average_rouge_scores['rougeL_precision']:.4f}")
print(f"ROUGE-L Recall: {average_rouge_scores['rougeL_recall']:.4f}")
print(f"ROUGE-L F-Measure: {average_rouge_scores['rougeL_fmeasure']:.4f}")

result_df.to_csv('/content/drive/MyDrive/DISSERTATION/Results/Relevance/Pegasus/peg_v4_rouge.csv', index=False)

### BART

bart_df=pd.read_csv("/content/drive/MyDrive/DISSERTATION/Dataset/test_bart_v2.csv")

scorer = rouge_scorer.RougeScorer(['rouge1', 'rouge2', 'rougeL'], use_stemmer=True)

#Function to calculate ROUGE scores
def calculate_rouge_scores(target, generated):
    scores = scorer.score(target, generated)
    return {
        'rouge1_precision': scores['rouge1'].precision,
        'rouge1_recall': scores['rouge1'].recall,
        'rouge1_fmeasure': scores['rouge1'].fmeasure,
        'rouge2_precision': scores['rouge2'].precision,
        'rouge2_recall': scores['rouge2'].recall,
        'rouge2_fmeasure': scores['rouge2'].fmeasure,
        'rougeL_precision': scores['rougeL'].precision,
        'rougeL_recall': scores['rougeL'].recall,
        'rougeL_fmeasure': scores['rougeL'].fmeasure
    }

rouge_scores = bart_df.apply(lambda row: calculate_rouge_scores(row['plain language summary'], row['generated_summary']), axis=1)


rouge_df = pd.DataFrame(rouge_scores.tolist())

result_df = pd.concat([bart_df, rouge_df], axis=1)

average_rouge_scores = rouge_df.mean()

#Average ROUGE scores
print("Average ROUGE Scores:")
print(f"ROUGE-1 Precision: {average_rouge_scores['rouge1_precision']:.4f}")
print(f"ROUGE-1 Recall: {average_rouge_scores['rouge1_recall']:.4f}")
print(f"ROUGE-1 F-Measure: {average_rouge_scores['rouge1_fmeasure']:.4f}")
print(f"ROUGE-2 Precision: {average_rouge_scores['rouge2_precision']:.4f}")
print(f"ROUGE-2 Recall: {average_rouge_scores['rouge2_recall']:.4f}")
print(f"ROUGE-2 F-Measure: {average_rouge_scores['rouge2_fmeasure']:.4f}")
print(f"ROUGE-L Precision: {average_rouge_scores['rougeL_precision']:.4f}")
print(f"ROUGE-L Recall: {average_rouge_scores['rougeL_recall']:.4f}")
print(f"ROUGE-L F-Measure: {average_rouge_scores['rougeL_fmeasure']:.4f}")

result_df.to_csv('/content/drive/MyDrive/DISSERTATION/Results/Relevance/BART/bart_v2_rouge.csv', index=False)

## BERTScore

pip install pandas bert-score

from bert_score import score

!pip install torch
import torch

### Pegasus

peg_df=pd.read_csv("/content/drive/MyDrive/DISSERTATION/Dataset/test_peg_v4.csv")

device = 'cuda' if torch.cuda.is_available() else 'cpu'

#Function to calculate BERTScore
def calculate_bertscore(targets, generations):
    P, R, F1 = score(generations, targets, lang='en', device=device, verbose=True)
    return P.tolist(), R.tolist(), F1.tolist()

precision, recall, f1 = calculate_bertscore(peg_df['plain language summary'].tolist(), peg_df['generated_summary'].tolist())

peg_df['bertscore_precision'] = precision
peg_df['bertscore_recall'] = recall
peg_df['bertscore_f1'] = f1

average_precision = peg_df['bertscore_precision'].mean()
average_recall = peg_df['bertscore_recall'].mean()
average_f1 = peg_df['bertscore_f1'].mean()

# Average BERTScores
print("Average BERTScores:")
print(f"Precision: {average_precision:.4f}")
print(f"Recall: {average_recall:.4f}")
print(f"F1: {average_f1:.4f}")

peg_df.to_csv('/content/drive/MyDrive/DISSERTATION/Results/Relevance/Pegasus/peg_v4_bertscore.csv', index=False)

### BART

bart_df=pd.read_csv("/content/drive/MyDrive/DISSERTATION/Dataset/test_bart_v2.csv")

device = 'cuda' if torch.cuda.is_available() else 'cpu'

#Function to calculate BERTScore
def calculate_bertscore(targets, generations):
    P, R, F1 = score(generations, targets, lang='en', device=device, verbose=True)
    return P.tolist(), R.tolist(), F1.tolist()

precision, recall, f1 = calculate_bertscore(bart_df['plain language summary'].tolist(), bart_df['generated_summary'].tolist())

bart_df['bertscore_precision'] = precision
bart_df['bertscore_recall'] = recall
bart_df['bertscore_f1'] = f1

average_precision = bart_df['bertscore_precision'].mean()
average_recall = bart_df['bertscore_recall'].mean()
average_f1 = bart_df['bertscore_f1'].mean()

# Average BERTScores
print("Average BERTScores:")
print(f"Precision: {average_precision:.4f}")
print(f"Recall: {average_recall:.4f}")
print(f"F1: {average_f1:.4f}")

bart_df.to_csv('/content/drive/MyDrive/DISSERTATION/Results/Relevance/BART/bart_v2_bertscore.csv', index=False)
