==> Python Notebooks/8_Factual_Consistency.ipynb <==
# Factual Consistency

from google.colab import drive
drive.mount('/content/drive')

pip install openai

import pandas as pd
from openai import OpenAI

test_df=pd.read_csv('/content/drive/MyDrive/DISSERTATION/Dataset/Extracted Dataset (csv)/test_plos.csv')

## Pegasus

peg_df=pd.read_csv('/content/drive/MyDrive/DISSERTATION/Dataset/test_peg_v4.csv')

#Replacing the abstact without stopwords to the original abstract for evaluation
peg_df = peg_df.drop(columns=["abstract"])
peg_df["abstract"] = test_df["abstract"]

### Summary Ranking

df=peg_df.copy()

#Summary Ranking Prompt
df['Message']="Decide which of the following summary is more consistent with the article. Note that consistency means all information in the summary is supported by the article.\nArticle: "+df['abstract']+"\nSummary A: "+df['plain language summary']+"\nSummary B: "+df['generated_summary']+"\nExplain your reasoning step by step then answer (A or B) the question: "

#API Key
client = OpenAI(api_key="PASTE YOUR API KEY HERE")

#GPT response generation
def get_response(message):
    stream = client.chat.completions.create(
        model="gpt-4o-mini",
        messages=[{"role": "user", "content": message}],
        stream=True,
    )
    response_content = ""
    for chunk in stream:
        if chunk.choices[0].delta.content is not None:
            response_content += chunk.choices[0].delta.content
    return response_content

df["Response"]=df["Message"].apply(get_response)

df.to_csv('/content/drive/MyDrive/DISSERTATION/Results/Factual Consistency/peg_v4_summaryRanking.csv', index=False)

### Consistency Rating

df_copy=peg_df.copy()

#Consistency Rating Prompt
df_copy['Message']="Score the following summary given the corresponding article with respect to consistency from 1 to 10. Note that consistency measures how much information included in the summary is present in the source article. 10 points indicate the summary contains only statements that are entailed by the source document.\nSummary: " + df_copy["generated_summary"] + "\nSource Article: " + df_copy["abstract"] + "\nMarks: "

print(df_copy['Message'][0])

#API Key
client = OpenAI(api_key="PASTE YOUR API KEY HERE")

#GPT response generation
def get_response(message):
    stream = client.chat.completions.create(
        model="gpt-4o-mini",
        messages=[{"role": "user", "content": message}],
        stream=True,
    )
    response_content = ""
    for chunk in stream:
        if chunk.choices[0].delta.content is not None:
            response_content += chunk.choices[0].delta.content
    return response_content

df_copy["Response"]=df_copy["Message"].apply(get_response)

df_copy.to_csv('/content/drive/MyDrive/DISSERTATION/Results/Factual Consistency/peg_v4_consistencyRating2.csv', index=False)

## BART

bart_df=pd.read_csv('/content/drive/MyDrive/DISSERTATION/Dataset/test_bart_v2.csv')

bart_df = bart_df.drop(columns=["abstract"])
bart_df["abstract"] = test_df["abstract"]

### Summary Ranking

bart_df

df=bart_df.copy()

#Summary Ranking API
df['Message']="Decide which of the following summary is more consistent with the article. Note that consistency means all information in the summary is supported by the article.\nArticle: "+df['abstract']+"\nSummary A: "+df['plain language summary']+"\nSummary B: "+df['generated_summary']+"\nExplain your reasoning step by step then answer (A or B) the question: "

#API Key
client = OpenAI(api_key="PASTE YOUR API KEY HERE")

#API Response Generation
def get_response(message):
    stream = client.chat.completions.create(
        model="gpt-4o-mini",
        messages=[{"role": "user", "content": message}],
        stream=True,
    )
    response_content = ""
    for chunk in stream:
        if chunk.choices[0].delta.content is not None:
            response_content += chunk.choices[0].delta.content
    return response_content

df["Response"]=df["Message"].apply(get_response)

df.to_csv('/content/drive/MyDrive/DISSERTATION/Results/Factual Consistency/bart_v2_summaryRanking.csv', index=False)

### Consistency Rating

df=bart_df.copy()

#Consistency Rating Prompt
df['Message']="Score the following summary given the corresponding article with respect to consistency from 1 to 10. Note that consistency measures how much information included in the summary is present in the source article. 10 points indicate the summary contains only statements that are entailed by the source document.\nSummary: " + df["generated_summary"] + "\nSource Article: " + df["abstract"] + "\nMarks: "

print(df['Message'][0])

#API Key
client = OpenAI(api_key="PASTE YOUR API KEY HERE")

#API Response Generation
def get_response(message):
    stream = client.chat.completions.create(
        model="gpt-4o-mini",
        messages=[{"role": "user", "content": message}],
        stream=True,
    )
    response_content = ""
    for chunk in stream:
        if chunk.choices[0].delta.content is not None:
            response_content += chunk.choices[0].delta.content
    return response_content

df["Response"]=df["Message"].apply(get_response)

df.to_csv('/content/drive/MyDrive/DISSERTATION/Results/Factual Consistency/bart_v2_consistencyRating2.csv', index=False)
