==> Python Notebooks/14 Extract Coherence Scores.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

import pandas as pd
import re

#Extracts the last decimal place
def extract_score_decimal(text):
    match = re.search(r'(?i)score[:\s]*\d+\.\d+', text)
    if match:
        return float(re.findall(r'\d+\.\d+', match.group())[0])
    return None

## Pegasus

peg_df=pd.read_csv('/content/drive/MyDrive/DISSERTATION/Results/Coherence/peg_v4_gpt2.csv')

peg_df['Answer'] = peg_df['Response'].apply(extract_score_decimal)

peg_df.to_csv('/content/drive/MyDrive/DISSERTATION/Results/Coherence/Extracted/peg_v4_gpt.csv', index=False)

print(f'Average Coherence score for Pegasus: {peg_df["Answer"].mean()}')

## BART

bart_df=pd.read_csv('/content/drive/MyDrive/DISSERTATION/Results/Coherence/bart_v2_gpt2.csv')

bart_df['Answer'] = bart_df['Response'].apply(extract_score_decimal)

bart_df.to_csv('/content/drive/MyDrive/DISSERTATION/Results/Coherence/Extracted/bart_v2_gpt.csv', index=False)

print(f'Average Coherence score for BART: {bart_df["Answer"].mean()}')
